\encoding{CP1252}
\name{IL}
\Rdversion{1.1}
\alias{IL}
\title{
A resampled Lax valuation dataset with 500 observations
}
\description{
This is a resampled version of the lax valuation dataset used for demonstrative
purposes only.
}
\usage{
data(IL)
}

\format{
  A data frame with 500 observations on the following variable
  \describe{
	  \item{\code{respondent no.}}{nr. of the self-selected intervals}
	  \item{\code{lower value}}{the left-end value of self-selected intervals}
	  \item{\code{upper value}}{the right-end value of self-selected intervals}
  }
}

\details{
This is a re-sampled version of the original CeHa (2009) data set.
}
\source{
see references
}
\references{
Belyaev, Yuri and Bengt Kriström. 2010. Approach to Analysis of Self-Selected Interval
Data. Technical report SLU, Department of Forest Economics 90183 Umeå, Sweden.
}

\keyword{datasets}
