\encoding{CP1252}
\name{iwtp-package}
\Rdversion{1.1}
\alias{iwtp-package}
\docType{package}
\title{
iwtp: Software for Analysis of Self-Selected Interval Data
}
\description{
This is a package for numerical evaluation of willingness to pay (WTP) based on interval data.
}
\details{
\tabular{ll}{
Package: \tab iwtp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-29\cr
License: \tab GPL-2\cr
LazyLoad: \tab no\cr
}

}
\author{
Yuri Belyaev, Wenchao Zhou

Maintainer: Wenchao Zhou <zhou.wenchao@sekon.slu.se>
}
\references{
Belyaev, Yuri and Bengt Kriström. 2010. Approach to Analysis of Self-Selected Interval
Data. Technical report SLU, Department of Forest Economics 90183 Umeå, Sweden.
}
\keyword{ package }
